requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.34
gTTS==2.3.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
protobuf==4.24.2
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
six==1.16.0
smmap==5.0.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
wavio==0.0.7
yarl==1.9.2
zipp==3.16.2
python-decouple 3.8
